+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.5.0] - 2021-11-09
+
 ### Added
 
 - Send email to Sales Admins and Customer Admin on organization and request updates.